refactor: separate dialog from the quiz itself
